improved simulated annealing to use the new creation of valid candidates, avoid to clone the current solution at each step and use slicing instead of torch.where to avoid intermediate tensor creations
