+++ thoi/heuristics/simulated_annealing.py
@@ -4,6 +4,7 @@ from tqdm import trange
 import torch
 
 from thoi.heuristics.scoring import _evaluate_nplets
+from thoi.heuristics.commons import _get_valid_candidates
 from thoi.commons import _normalize_input_data
 
 @torch.no_grad()
@@ -48,11 +49,8 @@ def simulated_annealing(X: Union[np.ndarray, torch.Tensor, List[np.ndarray], Lis
         current_energy = -current_energy
 
     # Initial valid candidates
-    all_elements = torch.arange(N, device=device)
-    valid_candidates = torch.stack([
-        all_elements[~torch.isin(all_elements, current_solution[b])]
-        for b in torch.arange(repeat, device=device)
-    ]).contiguous()
+    # |batch_size| x |N-order|
+    valid_candidates = _get_valid_candidates(current_solution, N, device)
 
     # Set initial temperature
     temp = initial_temp
@@ -62,6 +60,10 @@ def simulated_annealing(X: Union[np.ndarray, torch.Tensor, List[np.ndarray], Lis
     best_solution = current_solution.clone()
     # |batch_size|
     best_energy = current_energy.clone()
+    
+    # Repeat tensor for indexing the current_solution
+    # |repeat| x |1|
+    i_repeat = torch.arange(repeat)
 
     no_progress_count = 0
     pbar = trange(max_iterations, leave=False)
@@ -69,20 +71,21 @@ def simulated_annealing(X: Union[np.ndarray, torch.Tensor, List[np.ndarray], Lis
 
         pbar.set_description(f'mean({metric.upper()}) = {(1 if largest else -1) * best_energy.mean()} - ES: {no_progress_count}')
         
-        # Generate new solution by modifying the current solution
+        # Generate new solution by modifying the current solution.
+        # Generate the random indexes to change.
         # |batch_size| x |order|
-        new_solution = current_solution.clone()
-        i_sol = torch.randint(0, new_solution.shape[1], (repeat,), device=device)
+        i_sol = torch.randint(0, current_solution.shape[1], (repeat,), device=device)
         i_cand = torch.randint(0, valid_candidates.shape[1], (repeat,), device=device)
 
-        # Swap current values for candidates 
-        new_candidates = new_solution[torch.arange(repeat), i_sol]
-        current_candidates = valid_candidates[torch.arange(repeat), i_cand]
-        new_solution[torch.arange(repeat), i_sol] = current_candidates
+        # Update current values by new candidates and keep the original 
+        # candidates to restore where the new solution is not accepted.
+        current_candidates = current_solution[i_repeat, i_sol]
+        new_candidates = valid_candidates[i_repeat, i_cand]
+        current_solution[i_repeat, i_sol] = new_candidates
 
         # Calculate energy of new solution
         # |batch_size|
-        new_energy = _evaluate_nplets(covmats, T, new_solution, metric, use_cpu=use_cpu)
+        new_energy = _evaluate_nplets(covmats, T, current_solution, metric, use_cpu=use_cpu)
 
         if not largest:
             new_energy = -new_energy
@@ -98,43 +101,25 @@ def simulated_annealing(X: Union[np.ndarray, torch.Tensor, List[np.ndarray], Lis
         improves = delta_energy > 0
         accept_new_solution = torch.logical_and(improves, temp_probas)
         
-        # batch update solutions
+        # Restore original values for rejected candidates
         # |batch_size| x |order|
-        current_solution = torch.where(
-            accept_new_solution.unsqueeze(1).expand((-1,order)), # match correct shape
-            new_solution,
-            current_solution
-        )
+        current_solution[i_repeat[~accept_new_solution], i_sol[~accept_new_solution]] = current_candidates[~accept_new_solution]
+        
+        # Update valid_candidate for the accepted answers as they are not longer valid candidates
+        # |batch_size| x |N-order|
+        valid_candidates[i_repeat[accept_new_solution], i_cand[accept_new_solution]] = current_candidates[accept_new_solution]
 
+        # Update current energy for the accepted solutions
         # |batch_size|
-        current_energy = torch.where(
-            accept_new_solution,
-            new_energy,
-            current_energy
-        )
-
-        # update final valid_candidate depending on accepted answers
-        valid_candidates[torch.arange(repeat), i_cand] = torch.where(
-            accept_new_solution,
-            new_candidates,
-            current_candidates
-        )
+        current_energy[accept_new_solution] = new_energy[accept_new_solution]
 
         new_global_maxima = (new_energy > best_energy)
 
         # |batch_size| x |order|
-        best_solution = torch.where(
-            new_global_maxima.unsqueeze(1).expand((-1,order)), # match correct shape
-            new_solution,
-            best_solution
-        )
+        best_solution[new_global_maxima] = current_solution[new_global_maxima]
 
         # |batch_size|
-        best_energy = torch.where(
-            new_global_maxima,
-            new_energy,
-            best_energy
-        )
+        best_energy[new_global_maxima] = new_energy[new_global_maxima]
 
         # Cool down
         temp *= cooling_rate
